ecommerceBackend/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-restframework==0.0.1
pytz==2020.4
sqlparse==0.4.1
